+++ src/vesi/__init__.py
@@ -1,4 +1,4 @@
 """Vesi - Version control yang gampang dipelajari."""
 
-__version__ = "0.2.0"
+__version__ = "0.3.0"
 __codename__ = "vesi"